[peer]: update peer building strategy
