+++ peer-server/package.json
@@ -8,15 +8,15 @@
   "private": true,
   "scripts": {
     "dev": "ts-node --esm src/app.ts",
-    "build": "tsc",
+    "build": "esbuild src/app.ts --bundle --platform=node --outfile=./build/app.js",
     "test": "vitest run"
   },
   "dependencies": {
     "peer": "^1.0.1",
-    "winston": "^3.10.0",
-    "tools": "file:../tools"
+    "winston": "^3.10.0"
   },
   "devDependencies": {
+    "esbuild": "^0.19.2",
     "vitest": "^0.34.3",
     "ts-node": "^10.9.1",
     "websocket": "^1.0.34",
@@ -25,6 +25,5 @@
     "@types/node": "^20.5.7",
     "@types/express": "^4.17.17",
     "@types/websocket": "^1.0.6"
-  },
-  "type": "module"
+  }
 }
