+++ peer-server/test/ws.spec.ts
@@ -1,5 +1,5 @@
 import {describe, it, expect, beforeAll, afterAll} from 'vitest';
-import {main, stop} from '../src/app';
+import {main, stop} from '../src/bootstrap';
 import {utils} from 'tools';
 import {client as Websocket} from 'websocket';
 import fetch from 'node-fetch';
@@ -9,7 +9,7 @@ afterAll(() => stop());
 
 describe('Server test', () => {
     it('HTTP request for ID', () => new Promise<void>(done => {
-        fetch('http://localhost:7000/peerjs/id').then(res => {
+        fetch('http://localhost:7001/peerjs/id').then(res => {
             res.text().then(id => {
                 expect(id).toBeTypeOf('string');
                 expect(id).toMatch(/kuid-([0-9a-z]{4})-([0-9a-z]{4})-([0-9a-z]{4})-([0-9a-z]{4})/);
@@ -18,7 +18,7 @@ describe('Server test', () => {
         })
     }));
     it('Websocket connection can connect', () => new Promise<void>((resolve, reject) => {
-        const connectionURL = new URL('peerjs', 'ws://localhost:7000/');
+        const connectionURL = new URL('peerjs', 'ws://localhost:7001/');
         connectionURL.searchParams.set('key', 'peerjs');
         connectionURL.searchParams.set('id', utils.generateId('kuid'));
         connectionURL.searchParams.set('token', 'ni789lpa2d');
